feat: Add policy naming and improve proxy test setup
This commit introduces a `Name()` method to the `policy.Policy` interface and implements it for `imageManagementPolicy`, `podScopedPolicy`, and `readOnlyPolicy`. This allows policies to be identified by name.

Additionally, the `proxy.Server` now includes a `policyNames()` method to aggregate the names of applied policies, which is used in the `Version` response.

The `proxy_test.go` file has been added to provide unit tests for the proxy server, specifically for the `Version` method, using the `SetRuntimeClient` and `SetImageClient` methods for dependency injection.

Minor formatting adjustments were also made in `pkg/fake/server.go` and `test/e2e/e2e_test.go`.